+++ bin/params.dart
@@ -3,7 +3,7 @@ import 'dart:io' show Platform;
 import 'syslogger.dart' show Syslogger;
 
 final myname = (Platform.executable == 'dart') ? 'dmtr' : Platform.executable;
-final version = '0.1.31';
+final version = '0.1.32';
 String? optstr;
 String? addnote;
 bool pause = false;