+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+## 0.1.32
+- add extra messaging if there's wrong data in replies
+
 ## 0.1.31
-- add payload size changing at runtime
+- add payload size customization at runtime
 
 ## 0.1.30
 - add -s option (payload size)